fastapi/realtime/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==12.0
python-multipart==0.0.12
openai==1.57.0
python-dotenv==1.0.1
